30 June 2021
- just one thing

  project can be selected in Entry
  will be update Project Layer of the EditEntry

  how do i do it ?
  - at first I tried to parse project id from EntrySelectProjectFragment.kt to EditEntryFragment.kt
  then I get the project with projectId then observe the project

  -But this method does work but Project Layer of the EditEntry does not change even after selecting project. It only change when I click the save button.

 *** Problem was solved by directly updated projectId to database of Entry

+++ app/src/main/java/com/example/diaryapplication/database/ProjectDao.kt
@@ -16,7 +16,7 @@ interface ProjectDao{
     fun getAllProjects(): LiveData<List<Project>>
 
     @Query("SELECT * FROM project_table WHERE id = :id")
-    suspend fun getProject(id: Int): Project
+    suspend fun getProject(id: Long): Project
 
     @Query("SELECT * FROM project_table WHERE area = :area")
     fun getProjectsByArea(area : String): LiveData<List<Project>>
